+++ src/server/game/Entities/Player/Player.h
@@ -1211,7 +1211,7 @@ class Player : public Unit, public GridObject<Player>
         float m_anti_BeginFallZ; //alternative falling begin
         uint32 m_anti_lastalarmtime; //last time when alarm generated
         uint32 m_anti_alarmcount; //alarm counter
-        uint32 m_anti_TeleTime;
+        uint64 m_anti_TeleTime;
         uint32 m_PositionZ;
         bool m_CanFly;
         bool m_SpeedBuff;
@@ -1219,8 +1219,8 @@ class Player : public Unit, public GridObject<Player>
         bool m_FlyingMount;
         bool m_FlyingMount2;
         bool m_FlyingMountCheck;
-        uint32 Anti__GetLastTeleTime() const { return m_anti_TeleTime; }
-        void Anti__SetLastTeleTime(uint32 TeleTime) { m_anti_TeleTime=TeleTime; }
+        uint64 Anti__GetLastTeleTime() const { return m_anti_TeleTime; }
+        void Anti__SetLastTeleTime(uint64 TeleTime) { m_anti_TeleTime=TeleTime; }
         /*-------------------------------------*/
 
         void CleanupsBeforeDelete(bool finalCleanup = true);
